move SyncActions code to a separate file
